backend/requirements.txt
fastapi[standard]>=0.110.0
motor>=3.4.0
pydantic>=2.7.0
xgboost>=2.0.3
scikit-learn>=1.4.2
pandas>=2.2.2
numpy>=1.26.4
supabase>=2.3.0
python-dotenv>=1.0.0
